Replace ``logger.debug`` with a local ``call``
Instead of having ``noop`` be defined by commenting out invokes of
``subprocess.call``, a local function should be defined that can forward the
request if ``noop`` if False. Notice, for the time being, ``noop`` is
defaulting to True (or don't do anything). This script is still in its infancy;
I'm not ready for it (nor is the test environment) to start _actually_
calling/creating sub shells