Add asroot and noconfirm
makepkg complains because we are running as root. I leave it to you to trust if
it's an okay PKGBUILD; I already looked at it.

makepkg also calls pacman to get some extra packages we may not have already
installed; as such, because pacman is inherently interactive, we need to tell
it to just do the default action (similar to our `--noconfirms` above)